Added popup menu to group fragment
