+++ app/src/main/java/com/cornellappdev/android/pollo/GroupRecyclerView.kt
@@ -17,7 +17,7 @@ import kotlinx.coroutines.withContext
 
 class GroupRecyclerAdapter(
         private val groups: ArrayList<Group>,
-        val callback: GroupFragment.OnMoreButtonPressedListener?,
+        val callback: OnMoreButtonPressedListener?,
         val role: User.Role
 ) : androidx.recyclerview.widget.RecyclerView.Adapter<GroupRecyclerAdapter.ViewHolder>() {
 
@@ -112,6 +112,10 @@ class GroupRecyclerAdapter(
 
     internal fun getItem(index: Int): Group {
         return groups[index]
+    }
+
+    interface OnMoreButtonPressedListener {
+        fun onMoreButtonPressed(group: Group?)
     }
 
     companion object {
