+++ app/src/main/res/layout/fragment_main.xml
@@ -1,10 +1,14 @@
+<RelativeLayout
+    android:layout_height="match_parent"
+    android:layout_width="match_parent"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.cornellappdev.android.pollo.MainActivity$PlaceholderFragment">
+    android:layout_height="match_parent" >
 
 
     <FrameLayout
@@ -56,4 +60,15 @@
 
         </androidx.recyclerview.widget.RecyclerView>
     </FrameLayout>
+
+
 </androidx.constraintlayout.widget.ConstraintLayout>
+    <include
+        layout="@layout/manage_group_view"
+        android:id="@+id/groupMenuOptionsView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_alignBottom="@id/constraintLayout"
+        />
+
+</RelativeLayout>
