+++ app/src/main/res/layout/fragment_poll_group.xml
@@ -3,8 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.cornellappdev.android.pollo.MainActivity$PlaceholderFragment">
+    android:layout_height="match_parent">
 
 
     <FrameLayout
